Move filesystem functions to utils/fs
